feat(api/project): move delete_project_from_db to database module
